🚚 rename `show_connection` into `connections`
